Refactor consts
